requirements.txt
# Core dependencies
numpy>=1.20.0
mido>=1.2.10
pyyaml>=6.0
tqdm>=4.62.0  # Progress bars for Monte Carlo simulations
pandas>=1.3.0  # For data analysis in Monte Carlo simulations
matplotlib>=3.4.0  # For visualization of results
seaborn

# Optional dependencies
pytest>=6.2.5  # For running tests
sphinx>=4.0.0  # For generating documentation
jupyterlab>=3.2.0  # For interactive exploration of datasets